requirements.txt
crewai
exa-py
langchain
langchain-google-genai
streamlit
google-genai
exa_py
sendgrid
python-dotenv
-e. 
